requirements.txt
crewai>=0.28.0
crewai-tools>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
PyPDF2>=3.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
nltk>=3.8.1
sentence-transformers>=2.4.0
scikit-learn>=1.2.2